Renaming, modified semistructured constructor
