+++ examples/pysuccinct/pyfile.py
@@ -1,4 +1,4 @@
-import pyfile
+import file
 import sys
 import getopt
 
@@ -30,7 +30,7 @@ else:
     if (option[0] == "load"):
         # Load file from memory
         print("loading ", inputpath, " from file")
-        q = pyfile.PyFile(inputpath)
+        q = file.File(inputpath)
     elif (option[0] == "compress"):
         # Compress the file
         print("Please enter the sampling rates")
@@ -57,12 +57,14 @@ else:
             else:
                 printf("Invalid Option")
                 sys.exit(2)
-        q = pyfile.PyFile(inputpath, sa_sampling_rate, 
+        q = file.File(inputpath, sa_sampling_rate, 
             isa_sampling_rate, npa_sampling_rate, 
             sampling_scheme, npa_encoding_scheme)
     else:
         print("Usage: [load/compress] [file]\n")
         sys.exit(2)
+
+print("Command must be one of: search [query], count [query], extract [offset] [length]")
 
 # Parse through line by line
 while (True):
@@ -84,7 +86,7 @@ while (True):
         if (len(line) == 1):
             print("Could not parse command: ")
             continue
-        line = line[1].split(" ", 1)
+        line = line[1].split()
         offset = line[0].strip()
         length = line[1].strip()
         if (len(line) != 2 or RepresentsInt(offset) == False or RepresentsInt(length) == False):
