+++ examples/pysuccinct/pykv.py
@@ -1,4 +1,4 @@
-import pykv
+import kvstore
 import sys
 import getopt
 
@@ -30,7 +30,7 @@ else:
     if (option[0] == "load"):
         # Load file from memory
         print("loading ", inputpath, " from file")
-        q = pykv.PyKv(inputpath)
+        q = kvstore.Kvstore(inputpath)
     elif (option[0] == "compress"):
         # Compress the file
         print("Please enter the sampling rates")
@@ -57,7 +57,7 @@ else:
             else:
                 printf("Invalid Option")
                 sys.exit(2)
-        q = pykv.PyKv(inputpath, sa_sampling_rate, 
+        q = kvstore.Kvstore(inputpath, sa_sampling_rate, 
             isa_sampling_rate, npa_sampling_rate, 
             sampling_scheme, npa_encoding_scheme)
     else:
@@ -65,7 +65,6 @@ else:
         sys.exit(2)
 
 print("Command must be one of: search [query], count [query], get [key]")
-
 
 # Parse through line by line
 while (True):
