+++ pysuccinct/CMakeLists.txt
@@ -8,10 +8,10 @@ file(MAKE_DIRECTORY ${EXECUTABLE_OUTPUT_PATH})
 set(INCLUDE include ../core/include)
 include_directories(${INCLUDE})
 
-python_add_module(pyfile pyfile.cpp)
-python_add_module(pykv pykv.cpp)
-python_add_module(pysemistructured pysemistructured.cpp)
+python_add_module(file file.cpp)
+python_add_module(kvstore kvstore.cpp)
+python_add_module(semistructuredstore semistructuredstore.cpp)
 
-target_link_libraries(pyfile succinct)
-target_link_libraries(pykv succinct)
-target_link_libraries(pysemistructured succinct)
+target_link_libraries(file succinct)
+target_link_libraries(kvstore succinct)
+target_link_libraries(semistructuredstore succinct)
